Found symbolic expressions for H and Hdot in "Ball_Angular_Momentum" script. Added in some symbolic variables to do this. In all my scripts that run another script, I supressed printed outputs through a workaround that assigns anything that would normally be printed to the command line to a variable. Worked well. Maybe I should clear that variable after creating it.
Showed that my analytical method for finding Hdot was in fact correct by producing it in two ways, one straightforwardly differentiating H that was rotated into S0 and then rotating it back, and another by the analytical method I derived on paper. The difference between the two was the zero vector.