Moved the variable definitions script out of the kinematics folder and into the base directory for the repository - common to all the work. Started a plate kinetics file by copying the ball kinetics. Should be muc of the same stuff.
I began making the symVarVec and symFunVecs for use with the subs function but I need to change the names of the "kernal" symfuns (for x's and y's and thetas and such) so that I can perform the differentiation with those and then replace them with a symvar for display purposes. I'm going to use the " hat" notation for the symfuns. Need to make sure it doesnt screw anything up