Cleaned up the "change director" stuff so that it's on the user to make sure they are in the main directory before running any of the scripts. Found out that the code I used to set the directory didnt work on my laptop because my drives and directories are all different.
I think I am going to strip down some of the report-like elements of the Matlab live scripts and have a LaTeX document sort of running alongside the project - the report is going to get very large and I sort of want to be able to separate the work/thinking from the presentation. Might be too cumbersome to balance the two in the future.