Went away from the resample function and used interp1. There were weird artifacts of the resampling that popped up occassionally. These did not pop up when using interp1.
I decided to resample both signals at the same frequency because much of the literature has been suggesting that this is necessary. In this commit, I have the integral controller using two different sample frequencies for the DTW algorithm and the PID using the same frequency. I would like to understand why one method would be better than the other...it seems like the MATLAB documentation is suggestin the d_mn matrix doesnt need to be square and, intuitively, that makes sense to me. What I think happens is that the algorithm will "prefer" the longer signal because it has more points for points within the shorter signal to be close to. So, the shorter signal is more likely to be stretched to put  its data points close to the plethora of available points in the longer signal. The result is that the reference gets mostly preserved in the stretching.

This is tough because now I have another "dimension" of parameters to tweak in the comparison...the frequency ratio of the two signals AND the base frequency. I wish the literature would show to explore these issues but it doesnt.