In the very next script in the process, 'U_Joint_Kinematics', I implemented the Saved Workspaces change. Put in a lot of "Vdefs." to access the variables in the new way. Now I am debating if I should reorganize where my variables will live. Should all general, shared model variables/vectors/matrices/parameters live in one struct or should each step in the derivation process have its own relevant variables? Should I even be using structs at all? Maybe I should just load in the workspaces and access the data as I have been - this would make the reorganization a bit simpler, that's for sure. But it might be nice to have a bit of information in the context of a line of code, where the variable/data I'm using originates from. Though, this will produce longer lines of code and could clutter some things.
If I am going to use structs, I need to figure out how to handle that fact that a script may alter the state of a variable that was created in a different script.

Just thought of a solution: I can essentially make a rule for myself that any variables loaded in from external scripts should be considered read only. Then, if the script wants to assign some new data or meaning to a variable with the same name, it should do so "in its own workspace" in a sense, not alter the state of the variable within the struct that it loaded in. Then, if ANOTHER script wants to use the variable, it needs to essentially specify which version of it that it wants, the one that was created by the first script, or the one that was "altered" by the second script (there was no overriding but the second variable is the one that would have been altered in my old way of organizing). This way, the data loaded in from a particular script is in the state that it would be in when the script is finished running.

I have not implemented this change in the "U_Join_Kinematics" script yet but I will in the next commit. I will also alter the string that talks about the workspace data - it will reflect that the data in the saved .mat file should be considered read only.