Added in a few more changes to how variables are accessed. Realized that if I redefine gamma and beta symbolic variables it works around the bug I discovered earlier.
Realized in the middle of this whole reorganization process that it might be better to approach this whole project in an object oriented way. I can have the different steps in the derivation be classes. This way I can write scripts that don't depend on sharing data with other scripts. I made notes about this in onenote and began the design process for re-reorganizing the project. I think I am going to halt this process for now and branch AGAIN to try out the class approach. If the class approach doesnt pan out, I can always come back to this branch.