Added a MATLAB gitignore file that I found online.
Making some progress on the Var_Defs class. Input the state space model properties and changed some comments.

Decided to turn "Variable_Definitions" script into a sort of test script for the Var_Defs class. It will be a test run for the idea of using developement scripts for all future classes I write. The idea is that I can write code that documents the analysis behind what a class is doing in a way that links the livescript content to the actual content of the class .m file. I will use the class itself in the livescript, so each of the derivation steps taken in the old way will now be done using an object of the class. If I dont do this, I may want to make changes to the class or the current scripts and find that I have to try to keep the two up to date with each other. If they are tied together in this way, changing one affects the other intimately.