Played around with the gains some more. It's behaving much better now, I've tested various K matrices and beta stays within reasonable values for the circular setpoint with reasonable amplitude and frequency. Though the faster I make it, the worse the performance.
I was able to find a set of K matrices that made the integral controller actually perform better in the DTW measure than the PID controller. That raises the question - is the DTW measure really good for evaluating architectures, or is there so much performance variability within architectures that it can't be used for that purpose?