Fixed an issue where I wasnt accounting for the new input gain on Sdot that came up after I fixed the y_dot and y_ddot terms (v2bs a2bs).
After doing this, I believe the controller worked just fine but something I did changed that and it's not working anymore. I need to debug. I also found that the boundary layer gain and the switching gain are too intertwined with the size of the boundary layer. I think I am just going to write logic surrounding the size of the boundary layer and use a proportional controller on S inside and a maxed out (Tmax) signum control law outside the boundary layer.