Rearranged the Simulink files folder - now distinguishes between simulation files and hardware implementation files.
Got the Simulink Coder model to work. Readings are coming in and I can see them nicely when I'm, looking at a scope but the signal logging in the data inspector isn't working properly. I need to figure that out so I don't have scopes everywhere. Going to pick this up tomorrow.