Fixed a bug where u_FF wasn't a function of time if the feed-forward torque turned out to be a constant (as in the regulator case).
Decided I just want to get some HIL testing done and I'm going to do a "quick and dirty" solution before working out some of the issues I was encountering earlier. I started the file for this - for now, I'm just going to have the controller blocks in the main model instead of using a model reference for the controller.